feat(labware-creator): autofill homogenousWells for tipRacks
