+++ labware-library/src/labware-creator/components/sections/CreateNewDefinition.tsx
@@ -3,8 +3,8 @@ import { useFormikContext } from 'formik'
 import cx from 'classnames'
 import { PrimaryButton } from '@opentrons/components'
 import { Dropdown } from '../../components/Dropdown'
-import { isEveryFieldHidden } from '../../utils'
-import { labwareTypeOptions } from '../../fields'
+import { isEveryFieldHidden, makeAutofillOnChange } from '../../utils'
+import { labwareTypeOptions, labwareTypeAutofills } from '../../fields'
 import { FormAlerts } from '../FormAlerts'
 import { SectionBody } from './SectionBody'
 
@@ -18,18 +18,47 @@ interface Props {
   onClick: () => void
 }
 
-const Content = (props: Props): JSX.Element => {
+export const CreateNewDefinition = (props: Props): JSX.Element | null => {
+  const fieldList: Array<keyof LabwareFields> = [
+    'labwareType',
+    'tubeRackInsertLoadName',
+    'aluminumBlockType',
+    'aluminumBlockChildType',
+  ]
   const {
     disabled,
-    labwareTypeChildFields,
     onClick,
     showDropDownOptions,
+    labwareTypeChildFields,
   } = props
-  return (
+  const {
+    values,
+    errors,
+    touched,
+    setValues,
+    setTouched,
+  } = useFormikContext<LabwareFields>()
+
+  if (isEveryFieldHidden(fieldList, values)) {
+    return null
+  }
+
+  const content = (
     <div className={styles.labware_type_fields}>
       {showDropDownOptions && (
         <>
-          <Dropdown name="labwareType" options={labwareTypeOptions} />
+          <Dropdown
+            name="labwareType"
+            options={labwareTypeOptions}
+            onValueChange={makeAutofillOnChange({
+              name: 'labwareType',
+              autofills: labwareTypeAutofills,
+              values,
+              touched,
+              setTouched,
+              setValues,
+            })}
+          />
           {labwareTypeChildFields}
         </>
       )}
@@ -43,20 +72,6 @@ const Content = (props: Props): JSX.Element => {
       </PrimaryButton>
     </div>
   )
-}
-
-export const CreateNewDefinition = (props: Props): JSX.Element | null => {
-  const fieldList: Array<keyof LabwareFields> = [
-    'labwareType',
-    'tubeRackInsertLoadName',
-    'aluminumBlockType',
-    'aluminumBlockChildType',
-  ]
-  const { values, errors, touched } = useFormikContext<LabwareFields>()
-
-  if (isEveryFieldHidden(fieldList, values)) {
-    return null
-  }
 
   return (
     <div className={styles.new_definition_section}>
@@ -68,7 +83,7 @@ export const CreateNewDefinition = (props: Props): JSX.Element | null => {
       >
         <>
           <FormAlerts touched={touched} errors={errors} fieldList={fieldList} />
-          <Content {...props} />
+          {content}
         </>
       </SectionBody>
     </div>
