+++ labware-library/src/labware-creator/formSelectors.ts
@@ -7,6 +7,7 @@ import {
   getImplicitAutofillValues,
   DISPLAY_VOLUME_UNITS,
   tubeRackAutofills,
+  labwareTypeAutofills,
 } from './fields'
 import type { LabwareFields } from './fields'
 // TODO(Ian, 2019-07-24): consolidate `tubeRackAutofills/aluminumBlockAutofills`-getting logic btw here and makeAutofillOnChange
@@ -18,6 +19,10 @@ export const _getIsAutofilled = (
   const isAutofilledByDefault = Object.keys(
     getImplicitAutofillValues(values)
   ).includes(name)
+
+  if (labwareType === 'tipRack' && name === 'homogeneousWells') {
+    return Object.keys(labwareTypeAutofills[labwareType] || {}).includes(name)
+  }
 
   if (labwareType === 'aluminumBlock' && aluminumBlockType != null) {
     return (
