+++ labware-library/src/labware-creator/utils/index.ts
@@ -1,2 +1,3 @@
-export { isEveryFieldHidden } from './isEveryFieldHidden'
 export { displayAsTube } from './displayAsTube'
+export { isEveryFieldHidden } from './isEveryFieldHidden'
+export { makeAutofillOnChange } from './makeAutofillOnChange'
